+++ src/main/java/apptive/devlog/documentation/domain/auth/AuthSignupDoc.java
@@ -1,7 +1,7 @@
-package apptive.devlog.documentation.auth;
+package apptive.devlog.documentation.domain.auth;
 
 import apptive.devlog.common.response.success.CommonResponse;
-import apptive.devlog.documentation.wrapper.UserSignupResponseWrapper;
+import apptive.devlog.documentation.domain.user.wrapper.UserSignupResponseWrapper;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
